ml/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.4.3
pandas==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
